Refactor display class
